Rollback module deps creation Fix resolving tasks: now it is logged only once
