+++ example/subproject/src/main.py
@@ -1,9 +1,8 @@
 import sys
 from kernel import app
 
-print(sys.path)
-
 if __name__ == '__main__':
     print("Hello World")
+    print(sys.path)
     bar = app.Bar("name")
     print(bar.name)
