+++ plugins/python/src/main/kotlin/io/paddle/plugin/python/extensions/Interpreter.kt
@@ -1,11 +1,15 @@
 package io.paddle.plugin.python.extensions
 
 import io.paddle.plugin.python.dependencies.PyInterpreter
+import io.paddle.plugin.python.hasPython
 import io.paddle.project.PaddleProject
+import io.paddle.project.extensions.route
+import io.paddle.tasks.Task
 import io.paddle.utils.config.ConfigurationView
 import io.paddle.utils.ext.Extendable
 import io.paddle.utils.hash.Hashable
 import io.paddle.utils.hash.hashable
+import kotlin.system.measureTimeMillis
 
 
 val PaddleProject.interpreter: Interpreter
@@ -13,7 +17,17 @@ val PaddleProject.interpreter: Interpreter
 
 class Interpreter(val project: PaddleProject, val pythonVersion: PyInterpreter.Version) : Hashable {
 
-    val resolved: PyInterpreter by lazy { PyInterpreter.find(pythonVersion, project) }
+    val resolved: PyInterpreter by lazy {
+        project.terminal.info("Resolving interpreter...")
+        val result: PyInterpreter
+        measureTimeMillis {
+            checkInterpreterCompatibility()
+            result = PyInterpreter.find(pythonVersion, project)
+        }.also {
+            project.terminal.info("Finished: $it ms")
+        }
+        result
+    }
 
     object Extension : PaddleProject.Extension<Interpreter> {
         override val key: Extendable.Key<Interpreter> = Extendable.Key()
@@ -34,5 +48,19 @@ class Interpreter(val project: PaddleProject, val pythonVersion: PyInterpreter.V
 
     override fun hash(): String {
         return pythonVersion.number.hashable().hash()
+    }
+
+    private fun checkInterpreterCompatibility() {
+        if (!project.parents.all { it.hasPython }) {
+            return
+        }
+        for (parent in project.parents) {
+            if (parent.interpreter.pythonVersion != project.interpreter.pythonVersion) {
+                throw Task.ActException(
+                    "${parent.interpreter.pythonVersion.fullName} from ${":" + parent.route.joinToString(":")} " +
+                        "is not compatible with ${project.interpreter.pythonVersion.fullName} from ${":" + project.route.joinToString(":")}"
+                )
+            }
+        }
     }
 }
