+++ plugins/python/src/main/kotlin/io/paddle/plugin/python/tasks/resolve/ResolveInterpreterTask.kt
@@ -1,13 +1,10 @@
 package io.paddle.plugin.python.tasks.resolve
 
 import io.paddle.plugin.python.extensions.interpreter
-import io.paddle.plugin.python.hasPython
 import io.paddle.plugin.python.tasks.PythonPluginTaskGroups
 import io.paddle.project.PaddleProject
-import io.paddle.project.extensions.route
 import io.paddle.tasks.Task
 import io.paddle.tasks.incremental.IncrementalTask
-import kotlin.system.measureTimeMillis
 
 class ResolveInterpreterTask(project: PaddleProject) : IncrementalTask(project) {
     override val id: String = "resolveInterpreter"
@@ -18,25 +15,6 @@ class ResolveInterpreterTask(project: PaddleProject) : IncrementalTask(project) 
         get() = project.subprojects.getAllTasksById(this.id)
 
     override fun act() {
-        project.terminal.info("Resolving interpreter...")
-        val duration = measureTimeMillis {
-            checkInterpreterCompatibility()
-            project.interpreter.resolved
-        }
-        project.terminal.info("Finished: ${duration}ms")
-    }
-
-    private fun checkInterpreterCompatibility() {
-        if (!project.parents.all { it.hasPython }) {
-            return
-        }
-        for (parent in project.parents) {
-            if (parent.interpreter.pythonVersion != project.interpreter.pythonVersion) {
-                throw ActException(
-                    "${parent.interpreter.pythonVersion.fullName} from ${":" + parent.route.joinToString(":")} " +
-                        "is not compatible with ${project.interpreter.pythonVersion.fullName} from ${":" + project.route.joinToString(":")}"
-                )
-            }
-        }
+        project.interpreter.resolved
     }
 }
