+++ plugins/python/src/main/kotlin/io/paddle/plugin/python/tasks/resolve/ResolveRepositoriesTask.kt
@@ -5,7 +5,6 @@ import io.paddle.plugin.python.tasks.PythonPluginTaskGroups
 import io.paddle.project.PaddleProject
 import io.paddle.tasks.Task
 import io.paddle.tasks.incremental.IncrementalTask
-import kotlin.system.measureTimeMillis
 
 class ResolveRepositoriesTask(project: PaddleProject) : IncrementalTask(project) {
     override val id: String = "resolveRepositories"
@@ -16,8 +15,6 @@ class ResolveRepositoriesTask(project: PaddleProject) : IncrementalTask(project)
         get() = project.subprojects.getAllTasksById(this.id)
 
     override fun act() {
-        project.terminal.info("Resolving and indexing repositories...")
-        val duration = measureTimeMillis { project.repositories.resolved }
-        project.terminal.info("Finished: ${duration}ms")
+        project.repositories.resolved
     }
 }
